Test cases updated

+++ demo/test_stats.py
@@ -13,7 +13,7 @@ from numcompute.stats import (
 
 def test_normal_case():
     # Normal Case
-    X = np.array([1, 2, 3, 4, 5])
+    X = np.array([1, 2, 3, 4, 5], dtype=float)
 
     assert np.isclose(mean(X), 3)
     assert np.isclose(median(X), 3)
@@ -26,7 +26,7 @@ def test_normal_case():
 
 def test_with_nan_values():
     # With NaN values
-    X_nan = np.array([1, 2, np.nan, 4, 5])
+    X_nan = np.array([1, 2, np.nan, 4, 5], dtype=float)
 
     assert np.isclose(mean(X_nan), 3)
     assert np.isclose(median(X_nan), 3)
@@ -37,7 +37,7 @@ def test_with_nan_values():
 def test_axis_case():
     # 2D Array (axis test)
     X_2d = np.array([[1, 2, 3],
-                     [4, 5, 6]])
+                     [4, 5, 6]], dtype=float)
 
     assert np.allclose(mean(X_2d, axis=0), [2.5, 3.5, 4.5])
     assert np.allclose(mean(X_2d, axis=1), [2, 5])
@@ -45,7 +45,7 @@ def test_axis_case():
 
 def test_invalid_axis():
     # Invalid axis
-    X = np.array([1, 2, 3])
+    X = np.array([1, 2, 3], dtype=float)
 
     with pytest.raises(ValueError):
         mean(X, axis=1)
@@ -65,7 +65,7 @@ def test_non_numeric():
 
 def test_invalid_quantile():
     # Edge Case: Invalid Quantile
-    X = np.array([1, 2, 3, 4, 5])
+    X = np.array([1, 2, 3, 4, 5], dtype=float)
 
     with pytest.raises(ValueError):
         quantiles(X, [-10, 50, 110])
@@ -73,7 +73,7 @@ def test_invalid_quantile():
 
 def test_histogram_case():
     # Histogram
-    X = np.array([1, 2, 2, 3, 4])
+    X = np.array([1, 2, 2, 3, 4], dtype=float)
 
     hist, bin_edges = histogram(X, bins=3)
 
@@ -83,7 +83,7 @@ def test_histogram_case():
 
 def test_invalid_histogram_bins():
     # Invalid histogram bins
-    X = np.array([1, 2, 3])
+    X = np.array([1, 2, 3], dtype=float)
 
     with pytest.raises(ValueError):
         histogram(X, bins=0)
@@ -91,7 +91,7 @@ def test_invalid_histogram_bins():
 
 def test_invalid_histogram_range():
     # Invalid histogram range
-    X = np.array([1, 2, 3])
+    X = np.array([1, 2, 3], dtype=float)
 
     with pytest.raises(ValueError):
         histogram(X, bins=3, range=(5, 1))
@@ -99,7 +99,7 @@ def test_invalid_histogram_range():
 
 def test_describe():
     # Describe
-    X = np.array([1, 2, 3, 4, 5])
+    X = np.array([1, 2, 3, 4, 5], dtype=float)
     result = describe(X)
 
     assert isinstance(result, dict)
